Use official client wrappers and only provide a custom user agent client
